split off vector type header
