+++ include/ck/utility/amd_inline_asm.hpp
@@ -5,7 +5,7 @@
 #define CK_AMD_INLINE_ASM_HPP
 
 #include "c_style_pointer_cast.hpp"
-#include "data_type.hpp"
+#include "dtype_vector.hpp"
 
 // TODO: deprecate all amd_assembly_outer_product_xxx
